requirements.txt
torch==1.13.1+cu116 
torchvision==0.14.1+cu116 
torchaudio==0.13.1 
--extra-index-url https://download.pytorch.org/whl/cu116
numpy==1.26.4
pandas==2.2.3
pillow==10.4.0
requests==2.32.3
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
pydantic-settings==2.9.1
opencv-python==4.10.0.84
tifffile==2024.9.20
einops==0.8.0
scipy==1.14.1
timm==1.0.9
loguru==0.7.2
python-multipart==0.0.17
jinja2==3.1.4